+++ src/main/java/com/kuronami/musicdiscmaker/register/ModMenus.java
@@ -3,7 +3,7 @@ package com.kuronami.musicdiscmaker.register;
 import java.util.function.Supplier;
 
 import com.kuronami.musicdiscmaker.MusicDiscMaker;
-import com.kuronami.musicdiscmaker.menu.EnhancedJukeboxMenu;
+import com.kuronami.musicdiscmaker.menu.GoldenJukeboxMenu;
 import com.kuronami.musicdiscmaker.menu.MusicDiscMakerMenu;
 
 import net.minecraft.core.registries.Registries;
@@ -20,8 +20,8 @@ public final class ModMenus {
     public static final Supplier<MenuType<MusicDiscMakerMenu>> MUSIC_DISC_MAKER =
             MENUS.register("music_disc_maker", () -> IMenuTypeExtension.create(MusicDiscMakerMenu::new));
 
-    public static final Supplier<MenuType<EnhancedJukeboxMenu>> ENHANCED_JUKEBOX =
-            MENUS.register("enhanced_jukebox", () -> IMenuTypeExtension.create(EnhancedJukeboxMenu::new));
+    public static final Supplier<MenuType<GoldenJukeboxMenu>> GOLDEN_JUKEBOX =
+            MENUS.register("golden_jukebox", () -> IMenuTypeExtension.create(GoldenJukeboxMenu::new));
 
     private ModMenus() {
     }